+++ src/pause/pause_badges.c
@@ -350,7 +350,7 @@ void pause_badges_draw_contents(MenuPanel* menu, s32 baseX, s32 baseY, s32 width
     s32 orbColorR, orbColorG, orbColorB;
     s32 orbColorR1, orbColorG1, orbColorB1;
     s32 badgeIconElement;
-    s32 msg;
+    intptr_t msg;
     s32 msgX, msgY;
     s32 msgOpacity;
     s32 maxBP;
